+++ src/components/Menus/Sidebar/InstructionsPanel/InstructionsPanel.jsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useRef, useMemo } from "react";
+import React, { useEffect, useRef, useMemo, useState } from "react";
 import SidebarPanel from "../SidebarPanel";
 import { useTranslation } from "react-i18next";
 import { useSelector } from "react-redux";
@@ -8,6 +8,7 @@ import "prismjs/plugins/highlight-keywords/prism-highlight-keywords.js";
 import "prismjs/plugins/line-highlight/prism-line-highlight";
 import "prismjs/plugins/line-numbers/prism-line-numbers.css";
 import "prismjs/plugins/line-highlight/prism-line-highlight.css";
+import { quizReadyEvent } from "../../../../events/WebComponentCustomEvents";
 
 const InstructionsPanel = () => {
   const steps = useSelector((state) => state.instructions.project?.steps);
@@ -18,6 +19,7 @@ const InstructionsPanel = () => {
   );
   const { t } = useTranslation();
   const stepContent = useRef();
+  const [quizReady, setQuizReady] = useState(false);
 
   const isQuiz = useMemo(() => {
     return !!quiz?.questionCount;
@@ -42,10 +44,18 @@ const InstructionsPanel = () => {
 
     if (isQuiz) {
       setStepContent(quiz.questions[quiz.currentQuestion]);
+      setQuizReady(true);
     } else if (steps[currentStepPosition]) {
       setStepContent(steps[currentStepPosition].content);
     }
   }, [steps, currentStepPosition, quiz, isQuiz]);
+
+  useEffect(() => {
+    if (quizReady) {
+      document.dispatchEvent(quizReadyEvent);
+      setQuizReady(false);
+    }
+  }, [quizReady]);
 
   return (
     <SidebarPanel